server/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-Cors==5.0.0
Flask-Limiter==3.5.0
Flask-Session==0.8.0
Flask-Caching==2.3.0  # Added for caching support
Flask-SocketIO==5.3.6  # Added for WebSocket support
psycopg2-binary==2.9.9  # PostgreSQL adapter
Werkzeug==3.0.4
python-dotenv==1.0.1  # For environment variables
pytest==8.3.3  # For testing
pytest-flask==1.3.0  # For Flask-specific testing
reportlab==4.2.2  # Updated for PDF generation
openpyxl==3.1.5  # Updated for Excel generation
google-auth-oauthlib==1.2.1  # For Google OAuth 2.0
google-auth-httplib2==0.2.0  # For Google API requests
requests==2.32.3  # For HTTP requests (dependency for google-auth-oauthlib)
google-api-python-client==2.149.0  # For Google API client (provides googleapiclient)
redis==5.0.8  # Pinned to latest for caching and rate limiting
eventlet==0.36.0  # Pinned to latest for WebSocket support